requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
awscli==1.19.112
backports.functools-lru-cache==1.5
botocore==1.20.112
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.1
docutils==0.15.2
dt==1.1.39
fastapi==0.63.0
filelock==3.0.12
greenlet==1.0.0
h11==0.12.0
html5lib==1.0.1
idna==2.10
importlib-metadata==2.1.1
importlib-resources==3.3.0
iniconfig==1.1.1
jmespath==0.10.0
olefile==0.46
packaging==20.9
pathlib2==2.3.5
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycryptodomex==3.6.1
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.19.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.5
s3transfer==0.4.2
scandir==1.10.0
singledispatch==3.4.0.3
six==1.16.0
SQLAlchemy==1.4.12
SQLAlchemy-Utils==0.37.9
starlette==0.13.6
toml==0.10.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
virtualenv==20.2.2
webencodings==0.5.1
zipp==1.2.0
